fix(rag): decode ts_rank score as f64 (FLOAT8), keep pgvector as f32 (FLOAT4) + bump v0.1.6
- semantic_search uses pgvector '<=>' which returns FLOAT4 -> f32 (correct).
- keyword_search uses ts_rank which returns FLOAT8 -> f64 (was incorrectly changed to f32 in v0.1.5).
- Bump workspace to 0.1.6 + compose pins.

+++ pacgate-ai/crates/pacgate-rag/src/lib.rs
@@ -221,8 +221,9 @@ impl RagStore {
             .into_iter()
             .map(|row| KbSearchResult {
                 content: row.get("content"),
-                // ts_rank returns FLOAT4 (real), not FLOAT8. Decode as f32.
-                score: row.get::<f32, _>("score"),
+                // ts_rank returns FLOAT8 (double precision), not FLOAT4. Decode as
+                // f64 to match the SQL type and avoid a ColumnDecode panic.
+                score: row.get::<f64, _>("score") as f32,
                 source_doc: row.get("doc_name"),
                 page: row.get::<Option<i32>, _>("page").map(|p| p as u32),
             })
